Initial commit for discovery of AWS networks
Whilst the code at this commit mostly works, there is a random bug
which causes networks to be mis-typed as private if they are in the
process of being created. This is because the function may be running as
routes are being added. They are not later cleaned up to match the
correct location although this may be the result of the server side apply
process.